[SwiftSyntax] Move `error` function to standalone Utils.py
