+++ utils/gyb_syntax_support/Classification.py
@@ -1,4 +1,4 @@
-from .Node import error
+from .Utils import error
 from .kinds import lowercase_first_word  # noqa: I201
