+++ utils/gyb_syntax_support/Utils.py
@@ -1,0 +1,6 @@
+import sys
+
+
+def error(msg):
+    print('error: ' + msg, file=sys.stderr)
+    sys.exit(-1)
